+++ packages/apollo-cache-inmemory/CHANGELOG.md
@@ -1,6 +1,7 @@
 # Change log
 
 ### vNEXT
+- Fix error when missing __typename field in result [PR #2225](https://github.com/apollographql/apollo-client/pull/2225)
 - Refactored type usage
 - Prevented logging on defered queries
 - Refactored internal store from apollo-client into own package